requirements.txt
# K8s Agentic Decision Pipeline Requirements

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI and UI
click>=8.1.0
rich>=13.7.0

# Kubernetes and Prometheus
kubernetes>=28.1.0
prometheus-api-client>=0.5.0
requests>=2.31.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.9.0
tokentap>=0.1.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.12.0
ruff>=0.1.9
mypy>=1.7.0
